requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
python-dotenv==1.0.1
pytz==2024.2
typing_extensions==4.12.2
yarl==1.11.1
